pyproject.toml
[project]
name = "learning-backdoors"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.25.0",
    "bitsandbytes>=0.49.1",
    "blobfile>=3.0.0",
    "datasets>=4.5.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.1",
    "peft>=0.18.1",
    "pip>=25.3",
    "polars>=1.37.1",
    "python-dotenv>=1.0.0",
    "torch>=2.10.0",
    "transformers>=4.57.6",
    "wandb>=0.17.0",
]

[tool.ruff]
line-length = 120
target-version = "py312"


[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

[tool.ruff.lint.per-file-ignores]
"judge_responses.py" = ["E501"] # LLM as judge prompts

[dependency-groups]
dev = [
    "prek>=0.3.0",
    "ruff>=0.14.14",
]
